rename & remove unnecessary code
