+++ src/shared/address-details/action-table.tsx
@@ -1,4 +1,3 @@
-// tslint:disable:no-any
 import { ColumnProps } from "antd/es/table";
 import notification from "antd/lib/notification";
 import Table from "antd/lib/table";
@@ -29,9 +28,8 @@ export function getActionColumns(): Array<ColumnProps<ActionInfo>> {
     },
     {
       title: t("block.timestamp"),
-      render(_: string, record: any, __: number): string {
-        return String(record.block.timestamp);
-        // return record.actHash
+      render(_: string, record: ActionInfo, __: number): string {
+        return get(record, "block.timestamp");
       }
     },
     {
